Fixed: Element/ElementId hashCode and equals
