+++ accumulo/graph/src/main/java/org/vertexium/accumulo/AccumuloElement.java
@@ -412,29 +412,6 @@ public abstract class AccumuloElement extends ElementBase implements Serializabl
     }
 
     @Override
-    public int hashCode() {
-        return getId().hashCode();
-    }
-
-    @Override
-    public String toString() {
-        if (this instanceof Edge) {
-            Edge edge = (Edge) this;
-            return getId() + ":[" + edge.getVertexId(Direction.OUT) + "-" + edge.getLabel() + "->" + edge.getVertexId(Direction.IN) + "]";
-        }
-        return getId();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj instanceof Element) {
-            Element objElem = (Element) obj;
-            return getId().equals(objElem.getId());
-        }
-        return super.equals(obj);
-    }
-
-    @Override
     public Authorizations getAuthorizations() {
         return authorizations;
     }
