+++ core/src/main/java/org/vertexium/ElementBase.java
@@ -244,4 +244,27 @@ public abstract class ElementBase implements Element {
         }
         return elementIdProperty;
     }
+
+    @Override
+    public int hashCode() {
+        return getId().hashCode();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj instanceof ElementId) {
+            ElementId objElementId = (ElementId) obj;
+            return getId().equals(objElementId.getId()) && getElementType().equals(objElementId.getElementType());
+        }
+        return super.equals(obj);
+    }
+
+    @Override
+    public String toString() {
+        if (this instanceof Edge) {
+            Edge edge = (Edge) this;
+            return getId() + ":[" + edge.getVertexId(Direction.OUT) + "-" + edge.getLabel() + "->" + edge.getVertexId(Direction.IN) + "]";
+        }
+        return getId();
+    }
 }
