+++ inmemory/src/main/java/org/vertexium/inmemory/InMemoryElement.java
@@ -431,29 +431,6 @@ public abstract class InMemoryElement<TElement extends InMemoryElement> extends 
         }
     }
 
-    @Override
-    public int hashCode() {
-        return getId().hashCode();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj instanceof Element) {
-            Element objElem = (Element) obj;
-            return getId().equals(objElem.getId());
-        }
-        return super.equals(obj);
-    }
-
-    @Override
-    public String toString() {
-        if (this instanceof Edge) {
-            Edge edge = (Edge) this;
-            return getId() + ":[" + edge.getVertexId(Direction.OUT) + "-" + edge.getLabel() + "->" + edge.getVertexId(Direction.IN) + "]";
-        }
-        return getId();
-    }
-
     public boolean canRead(Authorizations authorizations) {
         return inMemoryTableElement.canRead(getFetchHints(), authorizations);
     }
